+++ Makefile
@@ -9,17 +9,24 @@ KERNEL_DIRECTORY=linux-$(KERNEL_VERSION)
 KERNEL_ARCHIVE=$(KERNEL_DIRECTORY).tar.xz
 KERNEL_URL=https://cdn.kernel.org/pub/linux/kernel/v$(KERNEL_MAJOR_VERSION).x/$(KERNEL_ARCHIVE)
 
+# Should we use the default features for the binaries
 RUST_USE_DEFAULT_FEATURES=true
+
+# Should we build in debug or release mode
+RUST_DEBUG_BUILD=true
 
 # Add features to enable for each program (using cargo feature syntax)
 RUST_FEATURES=
-RUST_FEATURES+= init/verbose_debug
 
 CARGO_FLAGS=--all --target=$(TARGET)
 
 # Turn off default package features
 ifeq ($(RUST_USE_DEFAULT_FEATURES), false)
 	CARGO_FLAGS+= --no-default-features
+endif
+
+ifneq ($(RUST_DEBUG_BUILD), true)
+	CARGO_FLAGS+= --release
 endif
 
 ifneq ($(RUST_FEATURES),)
