+++ src/init/src/defaults.rs
@@ -1,0 +1,10 @@
+/// The default Ibis logo to be printed at the boot banner
+pub const DEFAULT_BANNER_LOGO: &'static str = r#" _____ _     _     
+|_   _| |   (_)    
+  | | | |__  _ ___ 
+  | | | '_ \| / __|
+ _| |_| |_) | \__ \
+ \___/|_.__/|_|___/"#;
+
+/// Set the defaults for the PATH variable we want to set up
+pub const DEFAULT_PATH: &'static str = "/sbin;/bin";
